Restructuring of Test
NOTE: Pipeline not yet fully implemented, WholePipelineTest is therefore out-commented
