+++ Tests/testFunctions.py
@@ -1,3 +1,6 @@
+from pathlib import Path
+import os
+
 def CheckIfImageIsPassed(image, pipelineStep):
     if image is None:
         result = "Image is empty!!"
@@ -5,3 +8,8 @@ def CheckIfImageIsPassed(image, pipelineStep):
         result = "Image is not empty!!"
   
     print(pipelineStep+': '+result)
+
+def GetPath():
+    ROOT_DIR = os.path.realpath(os.path.dirname(__file__)+'/..')
+    relativePath = Path('Tests/Images')
+    return ROOT_DIR / relativePath
